Updated 26/10/2021

+++ Combinatorics/Catalan.cpp
@@ -1,4 +1,4 @@
 // Requires modinv
-ll catalan(ll n) {
+int catalan(int n) {
     return fac[2*n] * modinv(fac[n+1]) % MOD * modinv(fac[n]) % MOD;
 }
