+++ Combinatorics/Modpow.cpp
@@ -1,5 +1,5 @@
-ll modpow(ll n, ll x) {
-    ll res = 1;
+int modpow(int n, int x) {
+    int res = 1;
     while (x) {
         if (x & 1) res = (res * n) % MOD;
         x >>= 1, n = (n * n) % MOD;
