+++ Combinatorics/Permutations.cpp
@@ -1,4 +1,4 @@
 // Requires fac and modinv
-ll perm(ll n, ll r) {
+int perm(int n, int r) {
     return fac[n] * modinv(fac[n-r]) % MOD;
 }
